+++ Volund/src/Input/Input.h
@@ -14,7 +14,7 @@ namespace Volund
 		bool IsMouseButtonHeld(uint16_t button) const;
 		bool IsMouseButtonPressed(uint16_t button);
 
-		uint32_t GetScrollPosition() const;
+		float GetScrollPosition() const;
 
 		IVec2 GetMousePosition() const;
 
@@ -25,7 +25,7 @@ namespace Volund
 	private:
 		IVec2 m_MousePosition;
 
-		uint32_t m_ScrollPosition;
+		float m_ScrollPosition;
 
 		bool m_MouseButtons[VOLUND_MOUSE_BUTTON_AMOUNT];
 		bool m_Keys[VOLUND_KEY_AMOUNT];
